ml-training/requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
pymongo==4.6.0
python-dotenv==1.0.0
tqdm==4.66.1

